Almost finished history
Activity has been renamed, a maximum size for the history has been set up in AppData and a new method allows you to clear the history (still needs to be linked to the button in the activity). Groups in history are now sorted by newest first. ActionBar has been disabled in this activity, it now looks like the rating activity.

What still needs to be done:
fix the tests for MainActivity which are broken because we added a mandatory group in the intent. It can be created in the setup but for an unknown reason, the tests for tabs still doesn't work. I'm on it but haven't pushed yet.

Disable the history button if we are not in the groupsActivity, or make sure that we leave the group in which we are.