Fixed notification bug where a new instance of MainActivity was created by the notification
This instance was immediately finished because the Intent used by the activity didn't contain the needed GROUP flag.
The solution consists of not allowing the activity being instanciated multiple times but bringing the activity back on top of the stack instead.

We still need to update the notification content with the song being played.

Additional minor bug fix: we now see our UID in the prefsActivity instead of the UUID string.