First step of atomic deletion of a playlist, makes less calls to AndroidDB
When a deletion of a playlist form GS DB fails, it no more tries to remove it
form Android DB. But when a deletion from Android DB fails, it is not
restored to the GS DB. So the atomicity is not completely fulfilled.

The playlists objects are created only once, at activity creation time. When
changes are made to the playlists' ArrayList, only the adapter is reset.

Other changes:
* Removes deprecated comments
* UnisonDBHelper: Adds is_checked field to playlist table