Solo player can play songs from the playlist.
The SoloMainActivity retrieves a playlist from the android database. The player
itself does not hold the playlist. The PL is stored in the SoloMainActivity,
since it will be shared with the SoloTracksFragment. This way, it will allow
modifing the tracks order on the fly, during lecture (but this is "musique
d'avenir").
All the lecture stuff (next/previous/etc.) is done inside the PL object.

AbstractPlayerFragment: explicits the both app modes (groups, solo) by using an
enum (solo mode doesn't use the mHistPointer, since the playlist is completely
known before lecture).

Misc:
* Removes moved files (old locations)
* Removes unused solo_player.xml
* Organizes imports