Ends up ListViews conversion to Fragments (beta tests are successfull)
Each the local and the remote playlists lists are functionnal. Creating a new
playlist, saving it to the device, remove it from device are working quiet well.
Removing form the remote list is buggy, to be fixed soon.

A "+" button is added in the action bar to allow creating new playlists.

Else, the code of the SoloPlaylists* is cleaned, a lot of commented code is
removed.