Refactorizes ArrayAdapters and list row layouts
PlaylistsAdapter and TracksAdapter are now subclasses of Uutils
(PlaylistsAdapter is shared between at least two classes), making these
adapters usable for any activity of the project.

With the idea to make it a little more explicit, every layout describing a row
of a list is prefixed by "listrow", such that every such listrow layouts are
naturally grouped by name.
The row list layouts outside the scope of the solo mode are kept for
compatibility reasons, but they're intended to be removed after the merge with
the other branches.

Deprecated layouts such as list_fragment and solo_playlists are removed, since
they're no more used.