Refactorizes SoloMainActivity and its fragments
* Removes unused onPlaylistInfoListener
* Changes visibility of AbstractPlayer.next() from private to default
* Overrides AbstractPlayer.next() in SoloPlayerFragment
* Highlighting the currently played track in SoloTracksFragment doesn't work
* Improves listrow_track: adds an image to the left of the row, to be used to
  highlight the played track

The getListView() count in SoloTracksFragment always returns 0, thus I can't
access the ListView children. I haven't figured out how to overcome this issue
for now, that's why the highlighting doesn't work.

BUG the played track in GroupsPlayer doesn't correspond to the displayed data.
This is probably due to an issue when fetching the track from the DB (wrong
id).