FIX empty ListView of tracks in GroupsRatings
The tracks are back in the ratings activity. There was an issue with the
comparator (bad instanceof check!?!).

If your tracks are not shown, it is also possible that your database is
empty. The trick is then to reduce the LibraryService.MIN_UPDATE_INTERVAL to a
small value, like 60s (default value is 10h), and restart your AVD.

Commit applies:
* Organized imports
* Formatted project