Reduces I/O for AbstractFragmentActivities
* SoloPlaylistsActivity
Since the playlists fetched from the server are only the one created by the
user, there is no need to refresh. The user only accesses his content, so if he
does nothing, nothing happens and there's nothing to refresh.
But there's always the possibility to request a manual refresh by clicking the
refresh button in the action bar.

* AbstractFragmentActivity
  * Auto refresh
A boolean can be set for activities that doesn't request an automated refresh.
  * Default refresh interval
Since the content fetched from the server doesn't change much over time, the
refresh interval is increased from 30s to 60s.
There's always the possibility to request a manual refresh by clicking the
refresh button in the action bar.