FIX display only user's playlists as local GS playlists & new DB table interface
* Assume a user U creates a playlist and stores it on his device. If another
user V logs in with his own account on the same device, he would see user U's
local playlists. This is what I want to prevent with this FIX.
Even if the playlists can be modified outside the app, I don't want users to
mess up others playlists within the app, mainly for stability and consistency
reasons.
I will have to modify the playlist table in the sql db in order to handle the
case several users share the same device by tracking each account's playlists.

* Introduces a private interface in UnisonDB that will be implemented by each
table handler. The table handling will be refactorized as follow:
- An internal class will be defined per table, implementing the previously
  defined interface.
The idea is to group the methods by tables, and thus simplifiy and clarify the
class organization.