FIX Notification class callback
When clicking on the notification message in the notification bar, the player
appears in the correct mode. This means, that if you're playing music in the
solo mode, clicking the notification will bring you back to the player from solo
mode, instead of bringing you back to the groups (previously).

Creates a new Const type for Intents, to be used when serialized classes are
passed through intents.

FIX potential refactoring with a hard coded class name comparison in
GroupsHistoryActivity. Uses serialized class instead.