Fix NullPointerException; avoids login if no data connection
Fixes the NullPointerException when checking if database is open.

Tests the network status before trying to log in. If wifi or mobile connection
is on, the login can process. Else, a toast is shown to ask the user to check
his connection settings.
In order to check the network status, the corresponding permission has been
added in the Manifest.